requirements/requirements.txt
Django==1.6.5
PySimpleSOAP==1.10
PyYAML==3.11
South==1.0.1
TwitterAPI==2.2.5
configparser==3.5.0b2
dj-database-url==0.3.0
django-bootstrap3==5.1.1
django-debug-toolbar==1.2.1
django-extensions==1.5.0
django-guardian==1.2.5
djrill==1.1.1
docopt==0.4.0
facebookads==2.2.3
googleads==2.2.0
mailchimp==2.0.9
mock==1.0.1
oauthlib==0.7.2
python-dateutil==2.4.0
python-openid==2.2.5
python-social-auth==0.1.26
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
selenium==2.43.0
six==1.9.0
sqlparse==0.1.14
suds-jurko==0.6